Remove obsolete files
All of the .pyc and .pyo files are removed, as well as all .svn folders.
Also, the dist folder was removed since it contained duplicates of the
files and folders contained in the root of the repository.

Most (if not all) of the bat files have been removed to be replaced with
something more cross-platform. The entire locale and map-pack folders
have been removed, since they appear to be unused and outdated.